Fixed choppy gun aim

+++ Assets/Scripts/Physics/MultiUpdateBody.cs
@@ -7,7 +7,7 @@ public abstract class MultiUpdateBody : MultiUpdateObject
   protected Vector3 nextPosition;
   protected Vector3 interpolatedPosition {
     get {
-      return Vector3.Lerp(previousPosition, nextPosition, frameRatio);
+      return Vector3.Lerp(previousPosition, nextPosition, (float)frameRatio);
     }
   }
 
@@ -15,7 +15,7 @@ public abstract class MultiUpdateBody : MultiUpdateObject
   protected Quaternion nextRotation;
   protected Quaternion interpolatedRotation {
     get {
-      return Quaternion.Slerp(previousRotation, nextRotation, frameRatio);
+      return Quaternion.Slerp(previousRotation, nextRotation, (float)frameRatio);
     }
   }
 
@@ -27,7 +27,7 @@ public abstract class MultiUpdateBody : MultiUpdateObject
     nextPosition = reqRigidbody.position;
   }
 
-  override protected void ResetMultiUpdate(float time) {
+  override protected void ResetMultiUpdate(double time) {
     base.ResetMultiUpdate(time);
     nextPosition = previousPosition = reqRigidbody.position;
     nextRotation = previousRotation = reqRigidbody.rotation;
